--- middlewares/auth.js
@@ -1,7 +1,0 @@
-const auth = (req, res, next)=>{
-    
-    req.session.userSession? next(): res.redirect("/login")   
-}
-module.exports={
-    auth,
-}
