+++ src/models/Clientes.js
@@ -4,17 +4,13 @@ const {
 } = require('sequelize');
 module.exports = (sequelize, DataTypes) => {
   class Clientes extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
+
     static associate(models) {
-      // define association here
+      Clientes.belongsToMany(Profissional, { through: 'Agendamento' });
     }
   };
 
-  
+
   Clientes.init({
     nome: DataTypes.STRING,
     sobrenome: DataTypes.STRING,
@@ -26,3 +22,5 @@ module.exports = (sequelize, DataTypes) => {
   });
   return Clientes;
 };
+
+
