+++ src/models/Servico.js
@@ -1,0 +1,20 @@
+'use strict';
+const {
+  Model
+} = require('sequelize');
+module.exports = (sequelize, DataTypes) => {
+  class Servico extends Model {
+
+    static associate(models) {
+
+    }
+  };
+  Servico.init({
+    nome: DataTypes.STRING,
+    preco: DataTypes.FLOAT
+  }, {
+    sequelize,
+    modelName: 'Servico',
+  });
+  return Servico;
+};
