+++ src/models/agendamento.js
@@ -1,0 +1,23 @@
+'use strict';
+const {
+  Model
+} = require('sequelize');
+module.exports = (sequelize, DataTypes) => {
+  class Agendamento extends Model {
+    /**
+     * Helper method for defining associations.
+     * This method is not a part of Sequelize lifecycle.
+     * The `models/index` file will call this method automatically.
+     */
+    static associate(models) {
+      // define association here
+    }
+  };
+  Agendamento.init({
+    Data_Horario: DataTypes.DATE
+  }, {
+    sequelize,
+    modelName: 'Agendamento',
+  });
+  return Agendamento;
+};
