+++ src/views/dash.ejs
@@ -1,0 +1,119 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-BmbxuPwQa2lc/FVzBcNJ7UAyJxM6wuqIj61tLrc4wSX0szH/Ev+nYRRuWlolflfl" crossorigin="anonymous">
+    <link rel="preconnect" href="https://fonts.gstatic.com">
+    <link href="https://fonts.googleapis.com/css2?family=Nunito:wght@300;800&display=swap" rel="stylesheet">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.5.0/css/font-awesome.min.css">
+    <link rel="stylesheet" href="/stylesheets/dash.css">
+    <title>DashBoard</title>
+</head>
+
+<body>
+    
+    <%- include('./partials/profile') %>
+
+
+    <div class="col-7 principal">
+        <div class="row teste">
+            <div class="col">
+                <img src="/images/barberShop.jpg" alt="barberShop" class="barbershop">
+            </div>
+            <div class="col">
+                <img src="/images/logotipo_barbershop.svg" alt="logoBarber" class="barberLogo">
+                <p class="welcome">
+                    A Barbershop é uma barbearia com mais de 12 anos de serviços prestados e pode contar com os melhores
+                    profissionais de São Paulo para te atender.
+                </p>
+                <p>
+                    Avenida Sapopemba, 1020 | São Paulo - SP | Atendimento das 9h às 21h.
+                </p>
+                <button type="button" class="btn btn-warning btn-plus">Entrar em contato</button>
+
+            </div>
+        </div>
+        <div class="row disponiveis">
+
+            <h2 class="prof">Profissionais Disponíveis</h2>
+
+        </div>
+        <div class="row justify-content-around profissionais">
+            <% for (let barbeiro of barbeiros) { %>
+            <div class="col-4 teste info-agenda">
+
+                <img src="<%= barbeiro.img %>" alt="perfil" class="fotoProfissionais">
+                <p class="h5"><%= barbeiro.nome %></p>
+                <p><%= barbeiro.especialidade %></p>
+                <button type="button" class="btn btn-warning input-agenda"> <a
+                        href="/dash/<%= barbeiro.nome %>">Verificar agenda</a></button>
+
+            </div>
+            <% } %>
+        </div>
+    </div>
+    <div class="col-3 agendamento">
+        <p class="h2">Meus Agendamentos</p>
+        <p class="h5">Amanhã</p>
+        <div class="row">
+            <div class="col lista-agendamento">
+                <p class="service">Corte Simples</p>
+                <p class="service">08:00-08:30 AM</p>
+                <div class="detalhe">
+                    <img src="/images/funcionarios/miguel-souza.png" alt="perfil" class="fotoAgendamento">
+                    <button type="button" class="btn btn-warning input-agendamento">Verificar agenda</button>
+                </div>
+            </div>
+        </div>
+        <div class="row">
+            <div class="col lista-agendamento">
+                <p class="service">Corte Simples</p>
+                <p class="service">08:00-08:30 AM</p>
+                <div class="detalhe">
+                    <img src="/images/funcionarios/sidney-costella.png" alt="perfil" class="fotoAgendamento">
+                    <button type="button" class="btn btn-warning input-agendamento">Verificar agenda</button>
+                </div>
+            </div>
+        </div>
+        <p class="h5">9 de março</p>
+        <div class="row">
+            <div class="col lista-agendamento">
+                <p class="service">Corte Simples</p>
+                <p class="service">08:00-08:30 AM</p>
+                <div class="detalhe">
+                    <img src="/images/funcionarios/josemir-guster.png" alt="perfil" class="fotoAgendamento">
+                    <button type="button" class="btn btn-warning input-agendamento">Verificar agenda</button>
+                </div>
+            </div>
+        </div>
+        <div class="row">
+            <div class="col lista-agendamento">
+                <p class="service">Corte Simples</p>
+                <p class="service">08:00-08:30 AM</p>
+                <div class="detalhe">
+                    <img src="/images/funcionarios/paty-smith.png" alt="perfil" class="fotoAgendamento">
+                    <button type="button" class="btn btn-warning input-agendamento">Verificar agenda</button>
+                </div>
+            </div>
+        </div>
+        
+            </div>
+
+        </div>
+
+    </div>
+    <%- include('./partials/footer') %>
+
+    </div>
+    <script src="https://kit.fontawesome.com/0a254eab35.js" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-b5kHyXgcpbZJO/tY9Ul7kGkf1S0CWuKcCD38l8YkeH8z8QjE0GmW1gYU5S9FOnJ0" crossorigin="anonymous">
+    </script>
+
+</body>
+
+</html>
