+++ src/views/login.ejs
@@ -1,0 +1,78 @@
+<!DOCTYPE html>
+<html lang="pt-br">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Tela Inicial</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-BmbxuPwQa2lc/FVzBcNJ7UAyJxM6wuqIj61tLrc4wSX0szH/Ev+nYRRuWlolflfl" crossorigin="anonymous">
+    <link rel="preconnect" href="https://fonts.gstatic.com">
+    <link href="https://fonts.googleapis.com/css2?family=Nunito:wght@300;800&display=swap" rel="stylesheet">
+    <link rel="stylesheet" href="/stylesheets/style.css">
+</head>
+
+<body>
+
+    <div class="container index">
+        <div class="row">
+            <div class="col-7 texto">
+
+                <h1>Seja bem-vindo!</h1> <br><br>
+
+                <img class="ilustra" src="/images/ilustracao.svg" alt="Illustração">
+
+
+            </div>
+            <div class="col-5 mw-100 div-form">
+                <div class="form form-login">
+                    <div class="cadastro">
+                        <span>Login</span>
+                    </div>
+                    <% if(typeof(msg)!='undefined' ) { %>
+
+                        <small>
+                            <%= msg %>
+                        </small>
+
+                        <% } %>
+
+
+                            <form action="/login" method="POST">
+                                <div class="row">
+
+                                    <small id="passwordHelpBlock" class="form-text text-muted descricao">
+                                        Endereço de e-mail
+                                    </small>
+                                    <input type="email" id="inputPassword5" class="form-control input-maior"
+                                        aria-describedby="passwordHelpBlock" name="email">
+                                    <small id="passwordHelpBlock" class="form-text text-muted descricao">
+                                        Senha
+                                    </small>
+                                    <input type="password" id="inputPassword5" class="form-control input-maior"
+                                        aria-describedby="passwordHelpBlock" name="senha">
+
+                                    <small id="passwordHelpBlock" class="form-text text-muted descricao">
+                                        Manter Logado <input type="checkbox" name='lembreLogado'>
+                                    </small>
+                                    <button type="submit" class="btn btn-warning input-maior">Efetuar login</button>
+                                    <img src="/images/Logo.svg" alt="logo do site" class="logo">
+                                </div>
+                            </form>
+
+                </div>
+
+            </div>
+
+        </div>
+
+    </div>
+
+
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-b5kHyXgcpbZJO/tY9Ul7kGkf1S0CWuKcCD38l8YkeH8z8QjE0GmW1gYU5S9FOnJ0"
+        crossorigin="anonymous"></script>
+</body>
+
+</html>
